pyproject.toml
[project]
name = "drone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ardrone>=0.2.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "pygame-ce>=2.5.6",
    "pyparrot>=1.5.20",
]
